+++ packages/outbox/src/consumers/jobs-outbox.consumer.ts
@@ -1,0 +1,4 @@
+import type { JobsOutboxEntity, JobsOutboxModel } from '@volontariapp/database';
+import { OutboxConsumer } from '@volontariapp/database';
+
+export class JobsOutboxConsumer extends OutboxConsumer<JobsOutboxModel, JobsOutboxEntity> {}