+++ packages/outbox/src/outbox/writer/event-queue.writer.ts
@@ -1,3 +1,0 @@
-import { EventQueueEntity, EventQueueModel, OutboxWriter } from '@volontariapp/database';
-
-export class EventQueueWriter extends OutboxWriter<EventQueueModel, EventQueueEntity> {}