+++ packages/outbox/src/test/utils/repositories/jobs-outbox-test.repository.ts
@@ -1,11 +1,12 @@
-import { BaseRepository, JobsOutboxEntity, JobsOutboxModel, type Repository } from '@volontariapp/database';
+import { BaseRepository, JobsOutboxEntity, JobsOutboxModel } from '@volontariapp/database';
+import type { Repository } from 'typeorm';
 
-export class TestJobsOutboxWriterRepository extends BaseRepository<
+export class TestJobsOutboxRepository extends BaseRepository<
   JobsOutboxModel,
   JobsOutboxEntity,
   string
 > {
   constructor(repository: Repository<JobsOutboxModel>) {
-    super(repository as never, JobsOutboxEntity, JobsOutboxModel);
+    super(repository, JobsOutboxEntity, JobsOutboxModel);
   }
 }