+++ packages/database/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.9.0
+
+### Minor Changes
+
+- added outbox extended repository and fixed changelog history
+
 ## 1.8.0
 
 ### Minor Changes
@@ -26,6 +32,18 @@
 
 - Updated dependencies []:
   - @volontariapp/logger@0.2.3
+
+## 1.5.0
+
+### Minor Changes
+
+- patch database update
+
+## 1.4.0
+
+### Minor Changes
+
+- fix: update
 
 ## 1.3.0