+++ packages/outbox/CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- Added outbox writers & consumers implementations
+
 ## 0.1.2
 
 ### Patch Changes
 
 - Updated dependencies []:
   - @volontariapp/errors@0.5.0
-  - @volontariapp/database@1.8.1
 
 ## 0.1.1
 
@@ -14,8 +19,6 @@
 
 - Updated dependencies []:
   - @volontariapp/database@1.8.0
-
-All notable changes to this project will be documented in this file.
 
 ## 0.1.0